requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.22
Werkzeug==2.0.3
python-docx==0.8.11